Fixed bugs and added example images

+++ 3Dscanner_mk2/TextureDisplay.h
@@ -31,8 +31,8 @@ private:
     int image_x;
     int image_y;
     double center_of_rotation;
-    int top_crop;
-    int bottom_crop;
+    double top_crop;
+    double bottom_crop;
     ModelManager *model;
 };
